Запрос на менторство, исправил логику, написал тесты
