+++ src/main/java/school/faang/user_service/controller/user/UserController.java
@@ -26,7 +26,7 @@ public class UserController {
 
     @GetMapping("/{userId}")
     public UserDto getUser(@PathVariable long userId) {
-        return userService.getUserDto(userId);
+        return userService.getUser(userId);
     }
 
     @PostMapping
