--- src/main/java/school/faang/user_service/dto/RejectionDto.java
@@ -1,14 +1,0 @@
-package school.faang.user_service.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class RejectionDto {
-    private String reason;
-}
