--- src/main/java/school/faang/user_service/dto/RequestFilterDto.java
@@ -1,18 +1,0 @@
-package school.faang.user_service.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import school.faang.user_service.entity.RequestStatus;
-
-@Builder
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-public class RequestFilterDto {
-    private String description;
-    private Long requesterId;
-    private Long receiverId;
-    private RequestStatus status;
-}
