+++ src/main/java/school/faang/user_service/filter/mentorshiprequest/AuthorRequestFilter.java
@@ -1,21 +1,21 @@
-package school.faang.user_service.helper.filters;
+package school.faang.user_service.filter.mentorshiprequest;
 
 import org.springframework.stereotype.Component;
-import school.faang.user_service.dto.RequestFilterDto;
+import school.faang.user_service.dto.mentorshiprequest.RequestFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
 
 import java.util.stream.Stream;
 
 @Component
-public class RecipientRequestFilter implements RequestFilter {
+public class AuthorRequestFilter implements RequestFilter {
 
     @Override
     public boolean isApplicable(RequestFilterDto filterDto) {
-        return filterDto.getRequesterId() != null;
+        return filterDto.getReceiverId() != null;
     }
 
     @Override
     public Stream<MentorshipRequest> apply(Stream<MentorshipRequest> menReqs, RequestFilterDto filterDto) {
-        return menReqs.filter(menReq -> menReq.getRequester().getId() == (filterDto.getRequesterId()));
+        return menReqs.filter(menReq -> menReq.getReceiver().getId().equals(filterDto.getReceiverId()));
     }
 }
