+++ src/main/java/school/faang/user_service/filter/mentorshiprequest/RecipientRequestFilter.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.helper.filters.mentorshiprequest;
+package school.faang.user_service.filter.mentorshiprequest;
 
 import org.springframework.stereotype.Component;
 import school.faang.user_service.dto.mentorshiprequest.RequestFilterDto;
@@ -7,15 +7,15 @@ import school.faang.user_service.entity.MentorshipRequest;
 import java.util.stream.Stream;
 
 @Component
-public class DescriptionFilter implements RequestFilter {
+public class RecipientRequestFilter implements RequestFilter {
 
     @Override
     public boolean isApplicable(RequestFilterDto filterDto) {
-        return filterDto.getDescription() != null && !filterDto.getDescription().isBlank();
+        return filterDto.getRequesterId() != null;
     }
 
     @Override
     public Stream<MentorshipRequest> apply(Stream<MentorshipRequest> menReqs, RequestFilterDto filterDto) {
-        return menReqs.filter(menReq -> menReq.getDescription().contains(filterDto.getDescription()));
+        return menReqs.filter(menReq -> menReq.getRequester().getId().equals(filterDto.getRequesterId()));
     }
 }
