--- src/main/java/school/faang/user_service/helper/filters/mentorshiprequest/RequestFilter.java
@@ -1,15 +1,0 @@
-package school.faang.user_service.helper.filters.mentorshiprequest;
-
-import org.springframework.stereotype.Component;
-import school.faang.user_service.dto.mentorshiprequest.RequestFilterDto;
-import school.faang.user_service.entity.MentorshipRequest;
-
-import java.util.stream.Stream;
-
-@Component
-public interface RequestFilter {
-
-    boolean isApplicable(RequestFilterDto filterDto);
-
-    Stream<MentorshipRequest> apply(Stream<MentorshipRequest> menReqs, RequestFilterDto filterDto);
-}
