--- src/main/java/school/faang/user_service/services/UserService.java
@@ -1,18 +1,0 @@
-package school.faang.user_service.services;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import school.faang.user_service.entity.User;
-import school.faang.user_service.repository.UserRepository;
-
-import java.util.List;
-
-@Service
-@RequiredArgsConstructor
-public class UserService {
-    private final UserRepository userRepository;
-
-    public List<User> getUsersById(List<Long> usersId) {
-        return userRepository.findAllById(usersId);
-    }
-}
