+++ src/test/java/school/faang/user_service/helper/mentorshiprequest/DescriptionFilterTest.java
@@ -1,4 +1,4 @@
-package school.faang.user_service.helper.filters.mentorshiprequest;
+package school.faang.user_service.helper.mentorshiprequest;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
@@ -8,6 +8,7 @@ import org.mockito.InjectMocks;
 import org.mockito.junit.jupiter.MockitoExtension;
 import school.faang.user_service.dto.mentorshiprequest.RequestFilterDto;
 import school.faang.user_service.entity.MentorshipRequest;
+import school.faang.user_service.filter.mentorshiprequest.DescriptionFilter;
 
 import java.util.stream.Stream;
